move enums to separate module
