+++ packages/calcite-design-tokens/src/types/extensions.d.ts
@@ -1,6 +1,6 @@
 import type { Config as SDConfig, PlatformConfig as SDPlatformConfig, TransformedToken } from "style-dictionary/types";
 import { SetRequired } from "type-fest";
-import { Platform } from "./interfaces.js";
+import { Platform } from "../build/utils/enums.js";
 
 // convenience type per https://styledictionary.com/reference/utils/tokens/#converttokendata
 export type FlattenedTransformedToken = SetRequired<TransformedToken, "key">;
