+++ packages/calcite-design-tokens/src/types/interfaces.d.ts
@@ -1,11 +1,4 @@
+import { Platform } from "../build/utils/enums.js";
+
 export type Stylesheet = Extract<Platform, "css" | "scss">;
 export type RegisterFn = () => void;
-export enum Platform {
-  css = "css",
-  scss = "scss",
-  es6 = "es6",
-
-  // internal
-  js = "js",
-  docs = "docs",
-}
